Enable bugprone-signed-char-misuse

+++ src/editmap.cpp
@@ -475,7 +475,7 @@ void editmap::uber_draw_ter( const catacurses::window &w, map *m )
     creature_tracker &creatures = get_creature_tracker();
     drawsq_params params = drawsq_params().center( center );
     for( const tripoint &p : tripoint_range<tripoint>( start, end ) ) {
-        int sym = game_map ? '%' : ' ';
+        uint8_t sym = game_map ? '%' : ' ';
         if( p.x >= 0 && p.x < msize && p.y >= 0 && p.y < msize ) {
             if( game_map ) {
                 Creature *critter = creatures.creature_at( p );
